fix(ecdf): name each curve of a hue-grouped ECDF from its own colour
`sns.ecdfplot(hue=...)` draws one curve per group, and py-maidr named them
from the legend by position. seaborn draws an ECDF's hue levels in the
**reverse** of its legend order, so with two groups every point of both
curves was announced under the wrong name (#582).

Measured with groups that cannot be confused -- `low` is 1-5, `high` is
100-500, no overlap:

    legend texts : ['low', 'high']
    line 0       : x in [100, 500]     <- the "high" data
    line 1       : x in [1, 5]         <- the "low" data

    announced z='low'   x in [100, 500]   wrong
    announced z='high'  x in [1, 5]       wrong

Nothing errors and no value is altered. Only the names are exchanged, which
is the reading a user is least able to catch: the numbers look right, and the
curve they belong to is the one thing the chart was drawn to compare.

The fix is the match `patch/kdeplot` already makes for a density and
`scatterplot.hue_groups` makes point by point: a legend swatch names the
**colour** it is drawn in, and the colour survives an order that position
does not. That colour matching moves out of `patch/kdeplot` into
`util/legend_names` so the line family can use it too; `histogram` and the
existing tests import it from `patch/kdeplot` by name, so it is re-exported
there rather than moved out from under them.

Position remains the fallback, so nothing that was named before stops being
named: `names_for` declines rather than guesses -- no legend, a line no
swatch claims, a swatch naming two lines, fewer than two lines to tell apart
-- and every decline falls through to the pairing that was there.

The rest of the hue family was measured and is unchanged: `kdeplot(hue=)`
still names the curve peaking at 2.8 "low" and the one peaking at 281 "high",
and `lineplot(hue=)`, `ax.legend(["A", "B"])`, `ax.legend(handles=[q, p])`
(#578) and a legend shorter than its axes (#575) all still answer as they
did.

Closes #582
